migrations/versions/3c7afaea2f18_initial_database_setup.py
"""Initial database setup

Revision ID: 3c7afaea2f18
Revises: 
Create Date: 2025-06-29 16:25:41.929671

"""
from typing import Sequence, Union

from alembic import op
import sqlalchemy as sa


# revision identifiers, used by Alembic.
revision: str = '3c7afaea2f18'
down_revision: Union[str, Sequence[str], None] = None
branch_labels: Union[str, Sequence[str], None] = None
depends_on: Union[str, Sequence[str], None] = None


def upgrade() -> None:
    """Upgrade schema."""
    pass


def downgrade() -> None:
    """Downgrade schema."""
    pass
